PR suggetions: unique spec exceptions
